+++ src/lib.rs
@@ -14,81 +14,15 @@ limitations under the License.
 
 */
 
-use std::fmt;
-use std::result;
-
 use log::{debug, info, trace, warn};
 use rusqlite::Connection;
 use rusqlite::NO_PARAMS;
 
+mod errors;
+
 #[cfg(test)]
 mod tests;
-
-/// Enum listing possible errors.
-#[derive(Debug, PartialEq)]
-#[allow(clippy::enum_variant_names)]
-#[non_exhaustive]
-pub enum Error {
-    /// Rusqlite error, query may indicate the attempted SQL query
-    RusqliteError { query: String, err: rusqlite::Error },
-    /// Error with the specified schema version
-    SpecifiedSchemaVersion(SchemaVersionError),
-}
-
-impl Error {
-    fn with_sql(e: rusqlite::Error, sql: &str) -> Error {
-        Error::RusqliteError {
-            query: String::from(sql),
-            err: e,
-        }
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        // TODO Format the error with fmt instead of debug
-        write!(f, "rusqlite_migrate error: {:?}", self)
-    }
-}
-
-impl std::error::Error for Error {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        match self {
-            Error::RusqliteError { query: _, err } => Some(err),
-            Error::SpecifiedSchemaVersion(e) => Some(e),
-        }
-    }
-}
-
-impl From<rusqlite::Error> for Error {
-    fn from(e: rusqlite::Error) -> Error {
-        Error::RusqliteError {
-            query: String::new(),
-            err: e,
-        }
-    }
-}
-
-/// Errors related to schema versions
-#[derive(Debug, PartialEq, Clone, Copy)]
-#[allow(clippy::enum_variant_names)]
-#[non_exhaustive]
-pub enum SchemaVersionError {
-    /// Attempts to migrate to a version lower than the version currently in
-    /// the database. This is currently not supported
-    MigrateToLowerNotSupported,
-}
-
-impl fmt::Display for SchemaVersionError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "Attempts to migrate to a version lower than the version currently in the database. This is currently not supported.")
-    }
-}
-
-impl std::error::Error for SchemaVersionError {}
-
-/// A typedef of the result returned by many methods.
-pub type Result<T, E = Error> = result::Result<T, E>;
+pub use errors::{Error, Result, SchemaVersionError};
 
 /// One migration
 #[derive(Debug, PartialEq, Clone)]
